Introduce type ScalarValue and rename function scalarDefaultValue
