+++ packages/protobuf-bench/README.md
@@ -10,5 +10,5 @@ server would usually do.
 
 | code generator      | bundle size             | minified               | compressed         |
 |---------------------|------------------------:|-----------------------:|-------------------:|
-| protobuf-es         | 97,101 b      | 41,461 b | 10,746 b |
+| protobuf-es         | 97,081 b      | 41,461 b | 10,754 b |
 | protobuf-javascript | 394,384 b  | 288,654 b | 45,122 b |
