+++ packages/protobuf/src/google/protobuf/wrappers_pb.ts
@@ -20,7 +20,7 @@ import type { PartialMessage, PlainMessage } from "../../message.js";
 import { Message } from "../../message.js";
 import { proto3 } from "../../proto3.js";
 import type { JsonReadOptions, JsonValue, JsonWriteOptions } from "../../json-format.js";
-import { ScalarType } from "../../field.js";
+import { ScalarType } from "../../scalar.js";
 import type { FieldList } from "../../field-list.js";
 import type { BinaryReadOptions } from "../../binary-format.js";
 import { protoInt64 } from "../../proto-int64.js";
